requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
python-dotenv==1.0.1
PyJWT==2.9.0
flask-sock==0.7.0
gunicorn==22.0.0
gevent==24.2.1
